Split out usePickerProps hook and wired up ComboBox (#2074)
