+++ packages/components/src/spectrum/collections.ts
@@ -1,10 +1,16 @@
 export {
   ActionBar,
   type SpectrumActionBarProps as ActionBarProps,
+  // ComboBox is exported from ComboBox.tsx as a custom DH component. Re-exporting
+  // the Spectrum props type for upstream consumers that need to compose prop types.
+  type SpectrumComboBoxProps,
   // ListBox - we aren't planning to support this component
   MenuTrigger,
   type SpectrumMenuTriggerProps as MenuTriggerProps,
   // TableView - we aren't planning to support this component
+  // Picker is exported from Picker.tsx as a custom DH component. Re-exporting
+  // the Spectrum props type for upstream consumers that need to compose prop types.
+  type SpectrumPickerProps,
   TagGroup,
   type SpectrumTagGroupProps as TagGroupProps,
 } from '@adobe/react-spectrum';
