+++ packages/jsapi-components/src/spectrum/index.ts
@@ -1,2 +1,4 @@
+export * from './ComboBox';
 export * from './ListView';
 export * from './Picker';
+export * from './PickerProps';
