+++ packages/jsapi-components/src/spectrum/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './itemUtils';
 export * from './useItemRowDeserializer';
+export * from './usePickerProps';
